CHANGED: Created states to keep track of spindle work.
Created .c file implementing a way to get and set the states of spindle.
Added logic to prevent to turn spindle CCW when it is CW or vice-versa.
Fixed M6 while condition. This will prevent the M6 command to finish its execution.

TO DO: Make use of pantograph's sensors as a condition before end M3, M4 and M5 commands.