CHANGED: Exchanged board 1 and board 2 addresses, and updated IOs, and screen status accordingly.
This was an attempt to reduce disconnection problem of the boards. Refer to the link below for more details:
https://www.dynomotion.com/forum/viewtopic.php?f=14&t=1581&sid=da7c01c96248df284ab7f38190fcd055#p6099

What seemed to solve the problem was to short the distance from IO boards to the KFLOP control board, in order to reduce the noise injected when the spindle was switched on and off.